update espnetv2 format
