+++ configs/espnet/espnet_cityscapes_1024_512_120k_x2.yml
@@ -22,10 +22,8 @@ loss:
   coef: [1, 1]
 
 model:
-  type: EESPNetV2
-  backbone:
-    type: EESPNet_backbone
-    drop_prob: 0.0
-    scale: 2.0
+  type: ESPNetV2
+  in_channels: 3
+  scale: 2.0
   num_classes: 19
   drop_prob: 0.0
