+++ paddleseg/models/backbones/__init__.py
@@ -21,4 +21,3 @@ from .swin_transformer import *
 from .mobilenetv2 import *
 from .mix_transformer import *
 from .stdcnet import *
-from .espnet import *
