+++ paddleseg/models/layers/layer_libs.py
@@ -307,118 +307,3 @@ class ConvBNPReLU(nn.Layer):
         x = self._batch_norm(x)
         x = self._prelu(x)
         return x
-
-
-class BNPReLU(nn.Layer):
-    def __init__(self, out_channels, **kwargs):
-        super().__init__()
-        if 'data_format' in kwargs:
-            data_format = kwargs['data_format']
-        else:
-            data_format = 'NCHW'
-        self._batch_norm = SyncBatchNorm(out_channels, data_format=data_format)
-        self._prelu = layers.Activation("prelu")
-
-    def forward(self, x):
-        x = self._batch_norm(x)
-        x = self._prelu(x)
-        return x
-
-
-class EESP(nn.Layer):
-    """
-    EESP block, principle: reduce -> split -> transform -> merge
-
-    Args:
-        in_channels (int): Number of input channels.
-        out_channels (int): Number of output channels.
-        stride (int): Factor by which we should skip (useful for down-sampling). If 2, then down-samples the feature map by 2.
-        branches (int): Number of branches.
-        kernel_size_maximum (int): A maximum value of receptive field allowed for EESP block.
-        down_method (str): Down sample or not, only support 'avg' and 'esp'. (equivalent to stride is 2 or not)
-    """
-    def __init__(self,
-                 in_channels,
-                 out_channels,
-                 stride=1,
-                 branches=4,
-                 kernel_size_maximum=7,
-                 down_method='esp'):
-        super(EESP, self).__init__()
-        if out_channels % branches != 0:
-            raise RuntimeError(
-                "The out_channes for EESP should be factorized by branches, but out_channels={} cann't be factorized by branches={}"
-                .format(out_channels, branches))
-        assert down_method in [
-            'avg', 'esp'
-        ], "The down_method for EESP only support 'avg' or 'esp', but got down_method={}".format(
-            down_method)
-        self.in_channels = in_channels
-        self.stride = stride
-
-        in_branch_channels = int(out_channels / branches)
-        self.group_conv_in = ConvBNPReLU(in_channels,
-                                         in_branch_channels,
-                                         1,
-                                         stride=1,
-                                         groups=branches,
-                                         bias_attr=False)
-
-        map_ksize_dilation = {
-            3: 1,
-            5: 2,
-            7: 3,
-            9: 4,
-            11: 5,
-            13: 6,
-            15: 7,
-            17: 8
-        }
-        self.kernel_sizes = []
-        for i in range(branches):
-            kernel_size = 3 + 2 * i
-            kernel_size = kernel_size if kernel_size <= kernel_size_maximum else 3
-            self.kernel_sizes.append(kernel_size)
-        self.kernel_sizes.sort()
-
-        self.spp_modules = nn.LayerList()
-        for i in range(branches):
-            dilation = map_ksize_dilation[self.kernel_sizes[i]]
-            self.spp_modules.append(
-                nn.Conv2D(in_branch_channels,
-                          in_branch_channels,
-                          kernel_size=3,
-                          padding='same',
-                          stride=stride,
-                          dilation=dilation,
-                          groups=in_branch_channels,
-                          bias_attr=False))
-        self.group_conv_out = ConvBN(out_channels,
-                                     out_channels,
-                                     kernel_size=1,
-                                     stride=1,
-                                     groups=branches,
-                                     bias_attr=False)
-        self.bn_act = BNPReLU(out_channels)
-        self._act = nn.PReLU()
-        self.down_method = True if down_method == 'avg' else False
-
-    def forward(self, x):
-        group_out = self.group_conv_in(x)
-        output = [self.spp_modules[0](group_out)]
-
-        for k in range(1, len(self.spp_modules)):
-            output_k = self.spp_modules[k](group_out)
-            output_k = output_k + output[k - 1]
-            output.append(output_k)
-
-        group_merge = self.group_conv_out(
-            self.bn_act(paddle.concat(output, axis=1)))
-
-        if self.stride == 2 and self.down_method:
-            return group_merge
-
-        if group_merge.shape == x.shape:
-            group_merge = group_merge + x
-        out = self._act(group_merge)
-        return out
